Refactor settings and rag
