+++ src/paita/llm/services/service.py
@@ -4,11 +4,11 @@ from langchain_core.embeddings import Embeddings
 from langchain_core.language_models.chat_models import BaseChatModel
 
 from paita.llm.callbacks import AsyncHandler
-from paita.utils.settings_model import SettingsModel
+from paita.settings.llm_settings.llm_settings import LLMSettingsModel
 
 
 class Service:
-    def __init__(self, *, settings_model: SettingsModel, callback_handler: AsyncHandler):
+    def __init__(self, *, settings_model: LLMSettingsModel, callback_handler: AsyncHandler):
         self._settings_model = settings_model
         self._callback_handler = callback_handler
