+++ src/paita/tui/app.py
@@ -1,9 +1,7 @@
 from enum import Enum
 from pathlib import PurePath
-from typing import TYPE_CHECKING, Union
+from typing import Union, Optional
 
-from appdirs import user_config_dir
-from cache3 import DiskCache
 from textual.app import App, ComposeResult, Widget
 from textual.binding import Binding
 from textual.containers import Container, Horizontal, VerticalScroll
@@ -12,20 +10,16 @@ from textual.widgets import Button, Footer, Header, Input, LoadingIndicator
 from paita.llm.callbacks import AsyncHandler
 from paita.llm.chat import Chat
 from paita.llm.chat_history import ChatHistory
-from paita.llm.enums import Tag
-from paita.llm.models import list_all_models
 from paita.localization import labels
+from paita.settings.llm_settings.llm_settings import LLMSettings, LLMSettingsModel
+from paita.settings.rag_settings import RAG
 from paita.tui.error_screen import ErrorScreen
-from paita.tui.factory import create_rag_manager
 from paita.tui.message_box import MessageBox
 from paita.tui.multi_line_input import MultiLineInput
-from paita.tui.settings_screen import SettingsScreen
+from paita.tui.llm_settings_screen import LLMSettingsScreen
+from paita.tui.rag_settings_screen import RAGSettingsScreen
 from paita.tui.wait_screen import WaitScreen
 from paita.utils.logger import log
-from paita.utils.settings_manager import SettingsManager, SettingsModel
-
-if TYPE_CHECKING:
-    from paita.rag.rag_manager import RAGManager
 
 
 class Role(Enum):
@@ -33,11 +27,6 @@ class Role(Enum):
     ANSWER = "answer"
     INFO = "info"
     ERROR = "error"
-
-
-CACHE_DIR = user_config_dir(appname=labels.APP_TITLE, appauthor=labels.APP_AUTHOR)
-CACHE_NAME = "cache"
-CACHE_TTL = 24 * 60 * 60
 
 
 TEXT_AREA = True
@@ -51,16 +40,16 @@ class ChatApp(App):
     BINDINGS = [
         Binding("ctrl+q", "quit", "Quit", key_display="ctrl+q"),
         Binding("ctrl+x", "clear", "Clear", key_display="ctrl+x"),
-        Binding("ctrl+1", "settings", "Settings", key_display="ctrl+1"),
+        Binding("ctrl+1", "llm_settings", "LLM Settings", key_display="ctrl+1"),
+        Binding("ctrl+2", "rag_settings", "RAG Settings", key_display="ctrl+1"),
     ]
 
     def __init__(self):
         super().__init__()
 
-        self._settings_manager: SettingsManager = None
-        self._rag_manager: RAGManager = None
-        self._chat: Chat = None
-        self._cache: DiskCache = DiskCache(CACHE_DIR, CACHE_NAME)
+        self.settings: Optional[LLMSettings] = None
+        self.rag: Optional[RAG] = None
+        self._chat: Optional[Chat] = None
 
         self._current_message: Union[MessageBox or None] = None
         self._current_id: str = "id_0"
@@ -92,11 +81,21 @@ class ChatApp(App):
         if event.input.id in ("input", "multi_line_input"):
             await self.process_conversation()
 
-    def action_settings(self, allow_cancel: bool = False) -> None:  # noqa: FBT001, FBT002
-        settings_screen = SettingsScreen(
-            settings_manager=self._settings_manager,
-            rag_manager=self._rag_manager,
-            cache=self._cache,
+    def action_llm_settings(self, allow_cancel: bool = False) -> None:  # noqa: FBT001, FBT002
+        settings_screen = LLMSettingsScreen(
+            settings=self.settings,
+            allow_cancel=allow_cancel,
+        )
+        callback = self.exit_settings
+        log.info(f"{callback=}")
+        self.push_screen(
+            screen=settings_screen,
+            callback=callback,
+        )
+
+    def action_rag_settings(self, allow_cancel: bool = False) -> None:  # noqa: FBT001, FBT002
+        settings_screen = RAGSettingsScreen(
+            settings=self.settings,
             allow_cancel=allow_cancel,
         )
         callback = self.exit_settings
@@ -112,7 +111,7 @@ class ChatApp(App):
         await self.query_one("#body").mount(VerticalScroll(id="conversation"))
 
     def action_quit(self) -> None:
-        self._cache.clear()
+        self.settings.cache.clear()
         self.exit()
 
     # Callbacks
@@ -121,15 +120,34 @@ class ChatApp(App):
         if changed:
             self.init_chat()
 
-    async def on_mount(self) -> None:
+    def exit_rag_settings(self, changed: bool = False):  # noqa: FBT001, FBT002
+        if changed:
+            self.rag = RAG.create_rag(
+                app_name=labels.APP_TITLE,
+                app_author=labels.APP_AUTHOR,
+                settings_model=self.settings.settings_model,
+            )
+            self.init_chat()
+
+    async def on_mount(self):
         await self._mount_chat_history()
+        await self.push_screen(WaitScreen(labels.APP_LIST_AI_SERVICES_MODELS))
+
+        settings_exists = False
         try:
-            await self.push_screen(WaitScreen(labels.APP_LIST_AI_SERVICES_MODELS))
-            await self._list_models()
-            self.pop_screen()
+            self.settings: LLMSettings = await LLMSettings.load(app_name=labels.APP_TITLE, app_author=labels.APP_AUTHOR)
+            settings_exists = True
+        except FileNotFoundError:
+            self.settings = LLMSettings(
+                app_name=labels.APP_TITLE,
+                app_author=labels.APP_AUTHOR,
+            )
+
+        try:
+            await self.settings.refresh_llms()
         except ValueError as e:
             log.error(e)
-            self.pop_screen()
+            await self.pop_screen()
             await self.push_screen(
                 ErrorScreen(
                     labels.APP_ERROR_NO_AI_SERVICES_OR_MODELS,
@@ -138,43 +156,30 @@ class ChatApp(App):
                 self.exit,
             )
             return
-        # Read existing settings or open settings screen
-        try:
-            self._settings_manager: SettingsManager = await SettingsManager.load(
-                app_name=labels.APP_TITLE, app_author=labels.APP_AUTHOR
-            )
-            self._rag_manager = create_rag_manager(
+
+        await self.pop_screen()
+        if settings_exists:
+            self.rag = RAG.create_rag(
                 app_name=labels.APP_TITLE,
                 app_author=labels.APP_AUTHOR,
-                settings_model=self._settings_manager.model,
+                settings_model=self.settings.settings_model,
             )
-            await self._rag_manager.read()
             self.init_chat()
-        except FileNotFoundError:
-            self._settings_manager = SettingsManager(
-                cache=self._cache,
-                app_name=labels.APP_TITLE,
-                app_author=labels.APP_AUTHOR,
-            )
-            self._rag_manager = create_rag_manager(
-                app_name=labels.APP_TITLE,
-                app_author=labels.APP_AUTHOR,
-                settings_model=self._settings_manager.model,
-            )
-            await self._rag_manager.read()
+        else:
             self.action_settings(allow_cancel=False)
 
     def init_chat(self):
         if self._chat is None:
             self._chat = Chat()
+
         callback_handler = AsyncHandler()
         callback_handler.register_callbacks(self.callback_on_token, self.callback_on_end, self.callback_on_error)
 
         try:
             self._chat.init_model(
-                settings_model=self._settings_manager.model,
+                settings_model=self.settings.settings_model,
                 chat_history=self._chat_history,
-                rag_manager=self._rag_manager,
+                rag=self.rag,
                 callback_handler=callback_handler,
             )
             if TEXT_AREA:
@@ -182,16 +187,21 @@ class ChatApp(App):
             else:
                 self.query_one("#input").focus()
         except ValueError:
-            self._settings_manager = SettingsManager(
-                model=SettingsModel(),
+            self.settings = LLMSettings(
+                settings_model=LLMSettingsModel(),
                 app_name=labels.APP_TITLE,
                 app_author=labels.APP_AUTHOR,
+            )
+            self.rag = RAG.create_rag(
+                app_name=labels.APP_TITLE,
+                app_author=labels.APP_AUTHOR,
+                settings_model=self.settings.settings_model,
             )
             self.action_settings(allow_cancel=False)
 
     async def process_conversation(self) -> None:
         if TEXT_AREA:
-            text_input: MultiLineInput = self.query_one("#multi_line_input")
+            text_input: MultiLineInput = self.query_one("#multi_line_input", MultiLineInput)
             question = text_input.text
         else:
             text_input: Input = self.query_one("#input", Input)
@@ -222,11 +232,11 @@ class ChatApp(App):
         except ValueError as e:
             error = str(e)
             log.info(error)
-            self.push_screen(ErrorScreen(error), self.exit_error_screen)
+            await self.push_screen(ErrorScreen(error), self.exit_error_screen)
         except Exception as e:  # noqa: BLE001
             error = str(e)
             log.exception(e)
-            self.push_screen(ErrorScreen(error), self.exit_error_screen)
+            await self.push_screen(ErrorScreen(error), self.exit_error_screen)
 
     def toggle_widgets(self, *widgets: Widget) -> None:
         for w in widgets:
@@ -288,18 +298,6 @@ class ChatApp(App):
         conversation = self.query_one("#conversation")
         await conversation.mount_all(message_boxes)
         conversation.scroll_end(animate=False)
-
-    async def _list_models(self):
-        self.push_screen(WaitScreen(labels.APP_LIST_AI_SERVICES_MODELS))
-        all_models = await list_all_models()
-        self.pop_screen()
-        not_none_ai_models = any(len(models) for models in all_models.values())
-        if not not_none_ai_models:
-            msg = "No models found"
-            raise ValueError(msg)
-        for key in all_models:
-            if all_models[key]:
-                self._cache.set(key, all_models[key], CACHE_TTL, tag=Tag.AI_MODELS.value)
 
 
 def main():
