+++ src/paita/tui/styles/llm_settings_screen.tcss
